235-lowest-common-ancestor-of-a-binary-search-tree/src/Solution.java
public class Solution {

    public static TreeNode lowestCommonAncestor(TreeNode root, TreeNode p, TreeNode q) {
        if(root.val < p.val) {
            if(root.val < q.val) {
                return lowestCommonAncestor(root.right, p, q);
            } else {
                return root;
            }
        } else if(root.val > p.val){
            if(root.val > q.val) {
                return lowestCommonAncestor(root.left, p, q);
            } else {
                return root;
            }
        }
        return root;
    }

}
